Adds app-level scope to manage resource lifetimes:
ApplicationScope is a CoroutineScope but with application level fields like:
* a registry that holds resources to be freed/cleaned up before the app exits
* app properties from application.properties and local.properties available within the applicationScope for convenience

The main motivation for this change is to ensure proper cleanup of resources at the end of the app. Notably, I noticed SQLite connection being killed by the JVM becore it had a chance to commit all its transactions, and left things in a temporary journal file. Hopefully closing the db connection (and waiting for proper closure) before exiting the app ensure all transactions are commited.