Explicitly sets autocommit to true after migrations:
Liquibase will in fact alter the autocommit setting during its
execution. When Liquibase starts, it turns off autocommit mode to
ensure that all changes up to and including the database changelog lock
are performed within a single transaction. The concept behind this is
that if something fails during the application of changesets, the
transaction can be safely rolled back, leaving the database in a
consistent state.

However, one peculiarity of JDBC and most RDBMS implementations is that
once autocommit mode is turned off, it usually won't be turned back on
automatically. It's not directly Liquibase causing it, it's standard
JDBC behavior which Liquibase also adheres to.